readers-menu

+++ Menus/BookMenu.cpp
@@ -10,18 +10,17 @@
 
 
 void BookMenu::AddBook(BookService &b, GenreService &g) {
+    helpers.ClearScreen();
     std::string bookTitle = helpers.EnterData("Enter book title");
     std::string bookAuthor = helpers.EnterData("Enter book author");
-    int bookYearOfRelease = std::stoi(helpers.EnterData("Enter book year of release"));
+    int bookYearOfRelease = helpers.EnterInt("Enter book year of release");
     std::string bookLocationCode = helpers.EnterData("Enter book location");
-    int genreId;
 
     std::vector<Genre> genres = g.GetAllGenres();
     for (auto genre : genres) {
         std::cout << genre.Id << ". " << genre.Name << std::endl;
     }
-    std::cout << "Pick genre id " << std::endl;
-    std::cin >> genreId;
+    int genreId = helpers.EnterInt("Pick genre id");
 
     try {
         b.AddNewBook(bookTitle, bookAuthor, bookYearOfRelease, bookLocationCode, genreId);
@@ -29,34 +28,41 @@ void BookMenu::AddBook(BookService &b, GenreService &g) {
     } catch (std::exception &e) {
         std::cout << e.what() << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::GetBookById(BookService &b) {
-    int bookId = std::stoi(helpers.EnterData("Enter book id"));
+    helpers.ClearScreen();
+    int bookId = helpers.EnterInt("Enter book id");
     Book book = b.GetBook(bookId);
     if (book.Id <= 0) {
         std::cout << "There is no book with this id" << std::endl;
     } else {
         std::cout << book.Id << ". " << book.Title << ". " << book.Author << ". " << book.YearOfRelease << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::GetAllBooks(BookService &b) {
+    helpers.ClearScreen();
     std::vector<Book> books = b.GetAllBooks();
     for (auto book : books) {
         std::cout << book.Id << ". " << book.Title << ". " << book.Author << ". " << book.YearOfRelease << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::EditBook(BookService &b, GenreService &g) {
+    helpers.ClearScreen();
     std::cout << "Find book to edit" << std::endl;
     GetBookByName(b);
 
-    int bookId = std::stoi(helpers.EnterData("Enter book id"));
+    int bookId = helpers.EnterInt("Enter book id");
     Book book = b.GetBook(bookId);
 
     if (book.Id <= 0 || book.Title.empty()) {
         std::cout << "There is no book with this id" << std::endl;
+        helpers.PressEnterToContinue();
         return;
     }
 
@@ -68,16 +74,10 @@ void BookMenu::EditBook(BookService &b, GenreService &g) {
 
     bool isStillEditing = true;
     while (isStillEditing) {
-        std::cout << "Which one you want to edit:\n"
-                  << "1. Title\n"
-                  << "2. Author\n"
-                  << "3. Year of release\n"
-                  << "4. Location\n"
-                  << "5. Genre\n"
-                  << "6. Save\n"
-                  << "Choose: ";
-        int choice;
-        std::cin >> choice;
+        helpers.ClearScreen();
+        std::cout << "Editing: " << book.Title << std::endl;
+        std::cout << "1. Title\n2. Author\n3. Year of release\n4. Location\n5. Genre\n6. Save\n";
+        int choice = helpers.EnterInt("Choose");
 
         switch (choice) {
             case 1:
@@ -87,18 +87,18 @@ void BookMenu::EditBook(BookService &b, GenreService &g) {
                 bookAuthor = helpers.EnterData("Enter new author");
                 break;
             case 3:
-                bookYearOfRelease = std::stoi(helpers.EnterData("Enter new year of release"));
+                bookYearOfRelease = helpers.EnterInt("Enter new year of release");
                 break;
             case 4:
                 bookLocationCode = helpers.EnterData("Enter new location");
                 break;
             case 5: {
+                helpers.ClearScreen();
                 std::vector<Genre> genres = g.GetAllGenres();
                 for (auto genre : genres) {
                     std::cout << genre.Id << ". " << genre.Name << std::endl;
                 }
-                std::cout << "Pick genre id: ";
-                std::cin >> genreId;
+                genreId = helpers.EnterInt("Pick genre id");
                 break;
             }
             case 6:
@@ -115,17 +115,25 @@ void BookMenu::EditBook(BookService &b, GenreService &g) {
     } catch (std::exception &e) {
         std::cout << e.what() << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::DeleteBook(BookService &b) {
+    helpers.ClearScreen();
     GetBookByName(b);
-    int bookId = std::stoi(helpers.EnterData("Enter book id to DELETE"));
+    int bookId = helpers.EnterInt("Enter book id to DELETE");
+    if (!helpers.Confirm("Are you sure you want to delete this book?")) {
+        std::cout << "Cancelled" << std::endl;
+        helpers.PressEnterToContinue();
+        return;
+    }
     try {
         b.DeleteBook(bookId);
         std::cout << "Book deleted" << std::endl;
     } catch (std::exception &e) {
         std::cout << e.what() << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::GetBookByName(BookService &b) {
@@ -141,8 +149,9 @@ void BookMenu::GetBookByName(BookService &b) {
 }
 
 void BookMenu::AddBookCopy(BookService &b) {
+    helpers.ClearScreen();
     GetBookByName(b);
-    int bookId = std::stoi(helpers.EnterData("Enter book id"));
+    int bookId = helpers.EnterInt("Enter book id");
     std::cout << "Conditions: poor, average, good, new" << std::endl;
     std::string condition = helpers.EnterData("Enter condition");
     try {
@@ -151,22 +160,25 @@ void BookMenu::AddBookCopy(BookService &b) {
     } catch (std::exception &e) {
         std::cout << e.what() << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::EditBookCopy(BookService &b) {
+    helpers.ClearScreen();
     GetBookByName(b);
-    int bookId = std::stoi(helpers.EnterData("Enter book id to see copies"));
+    int bookId = helpers.EnterInt("Enter book id to see copies");
 
     std::vector<BookCopy> copies = b.FindBookCopies(bookId);
     if (copies.empty()) {
         std::cout << "No copies found" << std::endl;
+        helpers.PressEnterToContinue();
         return;
     }
     for (auto copy : copies) {
         std::cout << copy.Id << ". Condition: " << copy.Condition << std::endl;
     }
 
-    int copyId = std::stoi(helpers.EnterData("Enter copy id"));
+    int copyId = helpers.EnterInt("Enter copy id");
     std::cout << "Conditions: poor, average, good, new" << std::endl;
     std::string condition = helpers.EnterData("Enter new condition");
 
@@ -176,33 +188,43 @@ void BookMenu::EditBookCopy(BookService &b) {
     } catch (std::exception &e) {
         std::cout << e.what() << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::DeleteBookCopy(BookService &b) {
+    helpers.ClearScreen();
     GetBookByName(b);
-    int bookId = std::stoi(helpers.EnterData("Enter book id to see copies"));
+    int bookId = helpers.EnterInt("Enter book id to see copies");
 
     std::vector<BookCopy> copies = b.FindBookCopies(bookId);
     if (copies.empty()) {
         std::cout << "No copies found" << std::endl;
+        helpers.PressEnterToContinue();
         return;
     }
     for (auto copy : copies) {
         std::cout << copy.Id << ". Condition: " << copy.Condition << std::endl;
     }
 
-    int copyId = std::stoi(helpers.EnterData("Enter copy id to DELETE"));
+    int copyId = helpers.EnterInt("Enter copy id to DELETE");
+    if (!helpers.Confirm("Are you sure you want to delete this copy?")) {
+        std::cout << "Cancelled" << std::endl;
+        helpers.PressEnterToContinue();
+        return;
+    }
     try {
         b.DeleteBookCopy(copyId);
         std::cout << "Copy deleted" << std::endl;
     } catch (std::exception &e) {
         std::cout << e.what() << std::endl;
     }
+    helpers.PressEnterToContinue();
 }
 
 void BookMenu::Show(BookService &b, GenreService &g) {
     int choice;
     do {
+        helpers.ClearScreen();
         std::cout << "\n=== BOOK MENU ===\n";
         std::cout << "1. Books\n";
         std::cout << "2. Copies\n";
@@ -222,6 +244,7 @@ void BookMenu::Show(BookService &b, GenreService &g) {
 void BookMenu::ShowBooksMenu(BookService &b, GenreService &g) {
     int choice;
     do {
+        helpers.ClearScreen();
         std::cout << "\n=== BOOKS ===\n";
         std::cout << "1. Add book\n";
         std::cout << "2. Show all books\n";
@@ -237,7 +260,7 @@ void BookMenu::ShowBooksMenu(BookService &b, GenreService &g) {
             case 1: AddBook(b, g); break;
             case 2: GetAllBooks(b); break;
             case 3: GetBookById(b); break;
-            case 4: GetBookByName(b); break;
+            case 4: GetBookByName(b); helpers.PressEnterToContinue(); break;
             case 5: EditBook(b, g); break;
             case 6: DeleteBook(b); break;
             case 0: break;
@@ -249,6 +272,7 @@ void BookMenu::ShowBooksMenu(BookService &b, GenreService &g) {
 void BookMenu::ShowCopiesMenu(BookService &b) {
     int choice;
     do {
+        helpers.ClearScreen();
         std::cout << "\n=== COPIES ===\n";
         std::cout << "1. Add copy\n";
         std::cout << "2. Show copies of book\n";
@@ -261,12 +285,18 @@ void BookMenu::ShowCopiesMenu(BookService &b) {
         switch (choice) {
             case 1: AddBookCopy(b); break;
             case 2: {
+                helpers.ClearScreen();
                 GetBookByName(b);
                 int bookId = helpers.EnterInt("Enter book id");
                 auto copies = b.FindBookCopies(bookId);
-                for (auto copy : copies) {
-                    std::cout << copy.Id << ". Condition: " << copy.Condition << std::endl;
+                if (copies.empty()) {
+                    std::cout << "No copies found" << std::endl;
+                } else {
+                    for (auto copy : copies) {
+                        std::cout << copy.Id << ". Condition: " << copy.Condition << std::endl;
+                    }
                 }
+                helpers.PressEnterToContinue();
                 break;
             }
             case 3: EditBookCopy(b); break;
